Initial binary files
NXA-ENET8-POE_V1.1.2.5.bix is downloaded directly from my switch, the other two are from the OEM website. "Jump" is supposed to remap some flash or RAM and "Special" is supposed to allow for regular updates here out. ECS2100_V1.2.2.31.bix is the current firmware, and the target I am trying to run.